requirements.txt
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
oauth2==1.9.0.post1
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
seaborn==0.13.0
six==1.16.0
spotipy==2.23.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
